requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Bcrypt>=1.0.1
Pillow>=10.0.0
numpy>=1.24.0
pycryptodome>=3.20.0
reedsolo>=1.7.0
scipy>=1.11.0
Werkzeug>=3.0.0
gunicorn>=20.1.0
